Mock <DataBrokerImage> in tests
The images are wrapped in a <Suspense> call, but since those just
wait for images lazily loaded right after first render, rather than
after user input, it's a bit annoying to have to wait for that
loading in tests. Thus, I extracted it into a separate module and
mocked that module.
