+++ src/app/(proper_react)/redesign/(authenticated)/user/dashboard/Dashboard.test.tsx
@@ -29,6 +29,17 @@ jest.mock("next/navigation", () => ({
   usePathname: jest.fn(),
 }));
 
+jest.mock("../../../../../components/client/DataBrokerImage", () => {
+  return {
+    // Mock this with an empty React component. Otherwise, tests will complain:
+    // > Warning: A suspended resource finished loading inside a test, but the
+    // > event was not wrapped in act(...).
+    // > When testing, code that resolves suspended data should be wrapped into
+    // > act(...)
+    DataBrokerImage: () => null,
+  };
+});
+
 it("passes the axe accessibility test suite 1", async () => {
   const ComposedDashboard = composeStory(DashboardNonUsNoBreaches, Meta);
   const { container } = render(<ComposedDashboard />);
